+++ Editor/Source/Panel/PanelSystem.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "Panel.h"
+#include <vector>
+
+namespace Core
+{
+    class PanelSystem
+    {
+    private:
+        std::vector<Panel*> panels;
+
+    public:
+        /// @brief Will add default panels
+        PanelSystem();
+        ~PanelSystem();
+
+        void RenderImGui(PanelInformation *information);
+    };
+}