+++ Engine/Assets/Shaders/Object.vs.glsl
@@ -6,10 +6,11 @@ layout(location = 0) in vec3 aNormal;
 
 uniform mat4 uProjection;
 uniform mat4 uView;
+uniform mat4 uTransform;
 
 out vec2 vUVs;
 
 void main() {
     vUVs = aUVS;
-    gl_Position = uProjection * uView * vec4(aPosition, 1.0);
+    gl_Position = uProjection * uView * uTransform * vec4(aPosition, 1.0);
 }