+++ Engine/Source/Renderer/Renderer.h
@@ -45,5 +45,7 @@ namespace Core
         static void RenderScreenImage();
 
         static void Viewport(int width, int height);
+
+        static CeU32 GetPassID(int index);
     };
 }